+++ src/background-firefox.ts
@@ -8,9 +8,6 @@
 
 import { isJSONContentType } from "./content-type";
 
-// Keep track globally of URLs that contain JSON content.
-const jsonUrls = new Set<string>();
-
 function isRedirect(status: number) {
   return status >= 300 && status < 400;
 }
@@ -25,8 +22,10 @@ function detectJSON(event: chrome.webRequest.WebResponseHeadersDetails) {
       header.value &&
       isJSONContentType(header.value)
     ) {
-      jsonUrls.add(event.url);
+      addJsonUrl(event.url);
       if (typeof browser !== "undefined" && "filterResponseData" in browser.webRequest) {
+        // We need to change the content type to text/plain to prevent Firefox
+        // from using its built-in JSON viewer.
         header.value = "text/plain";
       }
     }
@@ -36,23 +35,39 @@ function detectJSON(event: chrome.webRequest.WebResponseHeadersDetails) {
 }
 
 // Listen for onHeaderReceived for the target page.
-// Set "blocking" and "responseHeaders".
 chrome.webRequest.onHeadersReceived.addListener(
   detectJSON,
+  // Firefox cannot fire onHeadersReceived for local files.
   { urls: ["<all_urls>"], types: ["main_frame"] },
   ["blocking", "responseHeaders"]
 );
 
-// Listen for a message from the content script to decide whether to operate on
-// the page. Calls sendResponse with a boolean that's true if the content script
-// should run, and false otherwise.
-chrome.runtime.onMessage.addListener((_message, sender, sendResponse) => {
-  if (sender.url?.startsWith("file://") && sender.url.endsWith(".json")) {
+chrome.runtime.onMessage.addListener((message, sender, sendResponse) => {
+  if (message !== "jsonview-is-json") {
+    return;
+  }
+
+  if (!sender.url) {
+    sendResponse(false);
+    return;
+  }
+
+  if (sender.url.startsWith("file://") && sender.url.endsWith(".json")) {
     sendResponse(true);
     return;
   }
-  sendResponse(sender.url && jsonUrls.has(sender.url));
-  if (sender.url) {
-    jsonUrls.delete(sender.url);
-  }
+
+  hasJsonUrl(sender.url).then(sendResponse);
+  return true; // this means "we're going to sendResponse asynchronously"
 });
+
+async function addJsonUrl(url: string) {
+  await chrome.storage.session.set({ [url]: true });
+}
+
+async function hasJsonUrl(url: string) {
+  const stored = await chrome.storage.session.get(url);
+  const present = url in stored;
+  await chrome.storage.session.remove(url);
+  return present;
+}