+++ src/content.ts
@@ -3,23 +3,11 @@ import { errorPage, jsonToHTML } from "./jsonformatter";
 import { installCollapseEventListeners } from "./collapse";
 import { safeStringEncodeNums } from "./safe-encode-numbers";
 
-function setJsonAsGlobalVariable(this: any, jsonObj: any) {
-  const script = document.createElement("script");
-  script.text = `Object.defineProperty(window, 'data', { value: ${JSON.stringify(
-    jsonObj
-  )}, writable: false, configurable: false });`;
-  document.documentElement.appendChild(script);
-
-  // log info message
-  // with this queueMicrotask user can not see source file information in log
-  queueMicrotask(() => console.log('JSON is exposed as a global variable called "data"'));
-}
-
 /**
  * This script runs on every page. It communicates with the background script
  * to help decide whether to treat the contents of the page as JSON.
  */
-chrome.runtime.sendMessage({}, (response: boolean) => {
+chrome.runtime.sendMessage("jsonview-is-json", (response: boolean) => {
   if (!response) {
     return;
   }
@@ -31,20 +19,21 @@ chrome.runtime.sendMessage({}, (response: boolean) => {
     content = jsonElems[0].textContent;
   } else {
     // Sometimes there's no pre? I'm not sure why this would happen
-    content = document.body.textContent;
+    content = (document.body.firstChild ?? document.body).textContent;
   }
   let outputDoc = "";
-  let jsonObj = null;
+  let jsonObj: any = null;
 
   if (content === null) {
     outputDoc = errorPage(new Error("No content"), "", document.URL);
   } else {
     try {
+      // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
       jsonObj = JSON.parse(safeStringEncodeNums(content));
       outputDoc = jsonToHTML(jsonObj, document.URL);
-    } catch (e: any) {
+    } catch (e) {
       outputDoc = errorPage(
-        e instanceof Error ? e : new Error(e.toString()),
+        e instanceof Error ? e : typeof e === "string" ? new Error(e) : new Error("Unknown error"),
         content,
         document.URL
       );
@@ -53,5 +42,4 @@ chrome.runtime.sendMessage({}, (response: boolean) => {
 
   document.documentElement.innerHTML = outputDoc;
   installCollapseEventListeners();
-  setJsonAsGlobalVariable(jsonObj);
 });